Student Management System
This projects is a simple GUI project made with PyQt and Sqlite3 for student database management system. In this you can add student details and payment details and show individual records associating a roll no.

Refer to the screenshots folder to have a general view of the working of the software.